speed up calculate_auc_values()
As mentioned in issue #15, the code for AUROC can be more efficient.

This commit improves the calculate_auc_values() function so that it runs
more quickly:

    Before: 1044 seconds for 145k cells, 28k genes, and 23 clusters
    After:   325 seconds

Before, the code was using parallel processing to repeat redundant
calculations independently for each cluster.

After, the code uses a single thread and does fewer redundant
calculations.

More details
------------

The underlying idea is that some intermediate values can be computed
once, saved, and then reused multiple times.

For example, it is necessary to sort and rank each gene to compute the
AUROC and the Wilcoxon rank sum test p-value.

It is inefficient to sort and rank each gene N times for N clusters.

A more efficient approach is to sort and rank each gene once. Then,
those ranks can be accessed multiple times to compute the statistics for
N clusters.